+++ src/main.rs
@@ -1,4 +1,4 @@
-use gtp;
+//use gtp;
 mod detached_command;
 
 fn main() {